// src/app/models/exchange/IExchangeSalesOrderListView.ts
export interface IExchangeSalesOrderListView {
    exchangeSalesOrderId: number;
    quoteExpireDate: string;
    versionNumber: string;
    customerId: number;
    customerName: string;
    customerCode: string;
    status: string;
    numberOfItems: number;
    createdDate: Date;
    updatedDate: Date;
    createdBy: string;
    updatedBy: string;
}